+++ lib/src/commands/fileupload.dart
@@ -1,11 +1,12 @@
 import 'dart:io';
 
-import 'package:ftpclient/src/debug/debuglog.dart';
 import 'package:path/path.dart';
 
 import '../ftpexceptions.dart';
 import '../ftpsocket.dart';
 import '../transfermode.dart';
+import '../debug/debuglog.dart';
+import '../util/transferutil.dart';
 
 class FileUpload {
   static const int _BUFFER_SIZE = 1024 * 1024;
@@ -21,20 +22,7 @@ class FileUpload {
     _log.log('Upload File: ${fFile.path}');
 
     // Transfer Mode
-    switch(_mode) {
-      case TransferMode.ascii:
-        // Set to ASCII mode
-        _socket.sendCommand('TYPE A');
-        _socket.readResponse();
-        break;
-      case TransferMode.binary:
-        // Set to BINARY mode
-        _socket.sendCommand('TYPE I');
-        _socket.readResponse();
-        break;
-      default:
-        break;
-    }
+    TransferUtil.setTransferMode(_socket, _mode);
 
     // Enter passive mode
     _socket.sendCommand('PASV');
@@ -44,7 +32,7 @@ class FileUpload {
       throw FTPException('Could not start Passive Mode', sResponse);
     }
 
-    int iPort = _parsePort(sResponse);
+    int iPort = TransferUtil.parsePort(sResponse);
 
     // Store File
     String sFilename = sRemoteName;
@@ -61,7 +49,7 @@ class FileUpload {
     RandomAccessFile fRAFile = fFile.openSync(mode: FileMode.read);
     int iRead = 0;
     final int iSize = fRAFile.lengthSync();
-    _log.log('File Size: $iSize');
+    _log.log('File Size: $iSize B');
 
     while (iRead < iSize) {
       int iEnd = _BUFFER_SIZE;
@@ -76,7 +64,7 @@ class FileUpload {
       iRead += iEnd;
     }
 
-    _log.log('Uploaded: $iRead');
+    _log.log('Uploaded: $iRead B');
 
     dataSocket.closeSync();
     fRAFile.closeSync();
@@ -84,19 +72,5 @@ class FileUpload {
     _socket.readResponse();
 
     _log.log('File Uploaded!');
-  }
-
-  /// Parse the Passive Mode Port from the Servers [sResponse]
-  int _parsePort(String sResponse) {
-    int iParOpen = sResponse.indexOf('(');
-    int iParClose = sResponse.indexOf(')');
-
-    String sParameters = sResponse.substring(iParOpen + 1, iParClose);
-    List<String> lstParameters = sParameters.split(',');
-
-    int iPort1 = int.parse(lstParameters[lstParameters.length - 2]);
-    int iPort2 = int.parse(lstParameters[lstParameters.length - 1]);
-
-    return (iPort1 * 256) + iPort2;
   }
 }