+++ lib/src/ftpexceptions.dart
@@ -3,4 +3,9 @@ class FTPException implements Exception {
   final String response;
 
   FTPException(this.message, [this.response]);
+
+  @override
+  String toString() {
+    return 'FTPException: $message (Response: $response)';
+  }
 }